requirements.txt
openai>=1.0.0
selenium>=4.0.0
requests>=2.31.0
python-dotenv==1.0.0
webdriver-manager>=4.0.0
fastapi>=0.110.0
uvicorn>=0.27.1
pydantic>=2.6.3
python-multipart>=0.0.9 